refactor: rename icons
